dashboard_claude/case-01-dashboard/requirements.txt
streamlit
psycopg2-binary
pandas
plotly
python-dotenv
